# Ensure runtime context on a per-device basis