Close #3320.

If we need a test for this, we'll need to run it as a subprocess, because somewhere else in the test suite we'll establish contexts for all devices, and this bug won't show up (which is why we didn't see it until now). Let me know if a test is desired.

I think the overhead of extra runtime API calls like `getDevice()` and `getDeviceCount()` are very low, so performance impact, if any, can only come from the usage of a Python list. I could try to do some quick benchmarks if desired, but I think this bug definitely needs to be fixed one way or another.