@leofang 2020-05-04
> I think the overhead of extra runtime API calls like `getDevice()` and `getDeviceCount()` are very low, so performance impact, if any, can only come from the usage of a Python list. I could try to do some quick benchmarks if desired, but I think this bug definitely needs to be fixed one way or another.

Added a snippet 
```python
cpdef test_ensure_context():
    _ensure_context()
```
to `cupy/cuda/runtime.pyx` and tested in ipython with `%timeit` magic, I got:

Before:
```bash
135 ns ± 0.386 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```
This PR (with Python list):
```bash                                                                                                               
193 ns ± 0.152 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```
Replace Python list by std::vector of bools:
```bash
179 ns ± 0.255 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```
If desired I can switch to the vector version.

(Note the timings are obtained after warm-up, since the first shot is much longer.)